Making trim_stories much faster by running a single query instead of n queries.

+++ apps/rss_feeds/models.py
@@ -164,12 +164,16 @@ class Feed(models.Model):
         user_stories_count = 0
         stories = Story.objects.filter(story_feed=self).order_by('-story_date')
         print 'Found %s stories in %s. Trimming...' % (stories.count(), self)
-        for story in stories[1000:]:
-            user_stories = UserStory.objects.filter(story=story)
-            user_stories_count = user_stories.count()
-            user_stories.delete()
-            story.delete()
-            stories_deleted_count += 1
+        if stories.count() > 1000:
+            for story in stories[1000:]:
+                user_stories = UserStory.objects.filter(story=story)
+                user_stories_count = user_stories.count()
+                user_stories.delete()
+                stories_deleted_count += 1
+            old_story = story[1000]
+            old_stories = Story.objects.filter(story_feed=self,
+                                               story_date__lte=old_story.story_date)
+            old_stories.delete()
         
         if stories_deleted_count:
             print "Trimming %s stories from %s. %s user stories." % (
